rename `err_with` to `error_with`
